backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend"
authors = ["Admin <admin@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
pyjwt = "^2.12.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
